+++ typescript_sdk/packages/reality-data-client/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @itwin/reality-data-client
 
-<!-- This log was last generated on Thu, 07 Nov 2024 14:16:10 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 08 Jul 2025 14:00:52 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.3.1
+
+Tue, 08 Jul 2025 14:00:52 GMT
+
+### Patches
+
+- Update core dev packages on devDependencies with released versions (bipin.pachhai@bentley.com)
+
+## 1.3.0
+
+Thu, 12 Jun 2025 14:09:49 GMT
+
+### Minor changes
+
+- Add support for iTwin.js Core 5.x (bipin.pachhai@bentley.com)
 
 ## 1.2.2